+++ packages/d1/src/components/Ui/ListItem/props.ts
@@ -7,6 +7,7 @@ import type { ProgressProps } from '../Progress'
 export const propsValues = {
   // :values [!] System label / Системная метка
   iconAlign: ['center', 'edge'],
+  controlPosition: ['start', 'end', 'autoSm', 'autoMd', 'autoLg', 'autoXl', 'auto2xl'],
   fill: [],
   size: ['sm', 'md', 'lg'],
   palette: ['red', 'orange', 'amber', 'yellow', 'lime', 'green', 'emerald', 'teal', 'cyan', 'sky', 'blue', 'indigo', 'violet', 'purple', 'fuchsia', 'pink', 'rose', 'slate', 'gray', 'zinc', 'neutral', 'stone', 'black', 'white']
@@ -17,12 +18,12 @@ type PropsToken = {
   // :type [!] System label / Системная метка
   focus?: boolean
   open?: boolean
-  selected?: boolean
   selectedChild?: boolean
   readonly?: boolean
   disabled?: boolean
   iconTop?: boolean
   iconAlign?: 'center' | 'edge'
+  controlPosition?: 'start' | 'end' | 'autoSm' | 'autoMd' | 'autoLg' | 'autoXl' | 'auto2xl'
   fill?: string | 'custom'
   divider?: boolean
   size?: 'sm' | 'md' | 'lg'
@@ -45,9 +46,11 @@ export type ListItemProps = ListItemPropsBasic<
  */
 export const defaults: object = {
   ...defaultsListItem,
+  iconCheckbox: 'check',
   ...{
     // :default [!] System label / Системная метка
     iconAlign: 'edge',
+    controlPosition: 'end',
     size: 'md'
     // :default [!] System label / Системная метка
   }